requirements.txt
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.11.0
matplotlib>=3.5.0
seaborn>=0.11.0
wordcloud>=1.8.0
streamlit>=1.25.0
snowflake-snowpark-python>=1.0.0
snowflake-ml-python>=1.0.0 